+++ rupring/src/lib.rs
@@ -1,10 +1,7 @@
-use std::collections::HashMap;
-
 pub mod boot;
 pub mod request;
+pub mod response;
 
-use http_body_util::Full;
-use hyper::{body::Bytes, header::HeaderName};
 pub use rupring_macro::{Controller, Delete, Get, Injectable, Module, Patch, Post, Put};
 
 extern crate rupring_macro;
@@ -12,29 +9,7 @@ extern crate rupring_macro;
 pub type Method = hyper::Method;
 
 pub use request::Request;
-
-#[derive(Debug, Clone, Default)]
-pub struct Response {
-    pub status: u16,
-    pub body: String,
-    pub headers: HashMap<HeaderName, String>,
-}
-
-impl From<Response> for hyper::Response<Full<Bytes>> {
-    fn from(response: Response) -> Self {
-        let mut builder = hyper::Response::builder();
-
-        builder = builder.status(response.status);
-
-        for (header_name, header_value) in response.headers {
-            builder = builder.header(header_name.clone(), header_value);
-        }
-
-        let response = builder.body(Full::new(Bytes::from(response.body))).unwrap();
-
-        return response;
-    }
-}
+pub use response::Response;
 
 pub trait IModule {
     fn child_modules(&self) -> Vec<Box<dyn IModule>>;
